# worthygoals-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "put": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update the authenticated user profile",
        "tags": [
          "User"
        ]
      }
    },
    "/users/me/data-export": {
      "post": {
        "operationId": "UsersController_exportMyData",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GDPR: export all data for the authenticated user",
        "tags": [
          "User"
        ]
      }
    },
    "/users/me": {
      "delete": {
        "operationId": "UsersController_deleteMyAccount",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GDPR: permanently delete the authenticated user account and all associated data",
        "tags": [
          "User"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create the profile for the authenticated account",
        "tags": [
          "User"
        ]
      }
    },
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "goalId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/complete": {
      "post": {
        "operationId": "TasksController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/explain": {
      "post": {
        "operationId": "TasksController_explain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Worthy Goals API",
    "description": "Backend Server API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email address of the user"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "The first name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "The last name of the user"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1990-01-01",
            "description": "The date of birth of the user"
          },
          "gender": {
            "type": "string",
            "example": "m",
            "description": "The gender of the user"
          },
          "latitude": {
            "type": "number",
            "example": 37.7749,
            "description": "The latitude of the user's location"
          },
          "longitude": {
            "type": "number",
            "example": -122.4194,
            "description": "The longitude of the user's location"
          },
          "tone": {
            "type": "string",
            "example": "firm",
            "description": "The user's onboarding tone preference (forced-choice deck result)"
          },
          "personalityId": {
            "type": "string",
            "example": "marcus",
            "description": "The user's onboarding-matched mentor (personality slug)"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Morning 5km run"
          },
          "goalId": {
            "type": "string",
            "example": "goal-uuid-here"
          },
          "description": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-06-01T07:00:00.000Z"
          },
          "repeatFrequency": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly",
              "monthly"
            ],
            "default": "none"
          }
        },
        "required": [
          "title",
          "goalId"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Morning 5km run"
          },
          "description": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-06-01T07:00:00.000Z"
          },
          "repeatFrequency": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly",
              "monthly"
            ],
            "default": "none"
          }
        }
      },
      "CompleteTaskDto": {
        "type": "object",
        "properties": {
          "moodScore": {
            "type": "number",
            "minimum": 1,
            "maximum": 4,
            "example": 3
          },
          "reflection": {
            "type": "string"
          },
          "memoryPictureId": {
            "type": "string",
            "description": "UUID of an uploaded media item (M06)"
          },
          "personalityId": {
            "type": "string",
            "description": "Personality ID for mentor reaction generation"
          }
        },
        "required": [
          "moodScore"
        ]
      },
      "ExplainTaskDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "couldnt",
              "forgot",
              "chose_not_to"
            ]
          },
          "freeText": {
            "type": "string"
          },
          "personalityId": {
            "type": "string",
            "description": "Personality ID for mentor reaction generation"
          }
        },
        "required": [
          "reason"
        ]
      },
...
    }
  }
}